08_middleware/pyproject.toml
[project]
name = "08-middleware"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain==1.0.0a5",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "langgraph-cli[inmem]>=0.4.2",
    "load-dotenv>=0.1.0",
]
